int vs size_tの部分を修正

+++ srcs/config/config.cpp
@@ -26,7 +26,7 @@ std::vector<Vserver> Config::GetServerVec() const { return server_vec_; }
 ConfigMap ConfigToMap(const Config &config) {
   ConfigMap config_map;
   std::vector<Vserver> server_vec = config.GetServerVec();
-  for (int i = 0; i < server_vec.size(); ++i) {
+  for (size_t i = 0; i < server_vec.size(); ++i) {
     config_map[server_vec[i].listen_].push_back(server_vec[i]);
   }
   return config_map;
@@ -37,18 +37,18 @@ std::ostream &operator<<(std::ostream &os, const Config &conf) {
   std::vector<Vserver> server_vec = conf.GetServerVec();
   os << "-------------------------------" << '\n';
 
-  for (int i = 0; i < server_vec.size(); ++i) {
+  for (size_t i = 0; i < server_vec.size(); ++i) {
     os << "server[" << i << "]:\n";
     os << "  listen: " << inet_ntoa(server_vec[i].listen_.sin_addr) << ":"
        << ntohs(server_vec[i].listen_.sin_port) << '\n';
     os << "  server_name: ";
-    for (int j = 0; j < server_vec[i].server_names_.size(); ++j) {
+    for (size_t j = 0; j < server_vec[i].server_names_.size(); ++j) {
       os << server_vec[i].server_names_[j] << " ";
     }
     os << '\n';
     os << "  timeout: " << server_vec[i].timeout_ << '\n';
     os << "  location: \n";
-    for (int j = 0; j < server_vec[i].locations_.size(); ++j) {
+    for (size_t j = 0; j < server_vec[i].locations_.size(); ++j) {
       os << "    locations_[" << j << "]:\n";
       os << "      match: " << server_vec[i].locations_[j].match_ << '\n';
       os << "      allow_method: ";
@@ -62,14 +62,14 @@ std::ostream &operator<<(std::ostream &os, const Config &conf) {
          << server_vec[i].locations_[j].client_max_body_size_ << '\n';
       os << "      root: " << server_vec[i].locations_[j].root_ << '\n';
       os << "      index: ";
-      for (int k = 0; k < server_vec[i].locations_[j].index_.size(); ++k) {
+      for (size_t k = 0; k < server_vec[i].locations_[j].index_.size(); ++k) {
         os << server_vec[i].locations_[j].index_[k] << " ";
       }
       os << '\n';
       os << "      is_cgi: " << server_vec[i].locations_[j].is_cgi_ << '\n';
       os << "      cgi_path: " << server_vec[i].locations_[j].cgi_path_ << '\n';
       os << "      error_pages: ";
-      for (int k = 0; k < server_vec[i].locations_[j].error_pages_.size();
+      for (size_t k = 0; k < server_vec[i].locations_[j].error_pages_.size();
            ++k) {
         os << server_vec[i].locations_[j].error_pages_[k] << " ";
       }
